Changed runtime files so that the fork makes more sense.
