+++ runtime/runtime.go
@@ -21,6 +21,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/pkg/errors"
+
 	"github.com/open-policy-agent/opa/ast"
 	"github.com/open-policy-agent/opa/loader"
 	"github.com/open-policy-agent/opa/plugins"
@@ -30,6 +32,7 @@ import (
 	"github.com/open-policy-agent/opa/repl"
 	"github.com/open-policy-agent/opa/server"
 	"github.com/open-policy-agent/opa/storage"
+	"github.com/open-policy-agent/opa/storage/inmem"
 	"github.com/open-policy-agent/opa/util"
 	"github.com/open-policy-agent/opa/version"
 )
@@ -153,6 +156,84 @@ type Runtime struct {
 	Manager *plugins.Manager
 
 	decisionLogger func(context.Context, *server.Info)
+}
+
+// NewRuntime returns a new Runtime object initialized with params.
+func NewRuntime(ctx context.Context, params Params) (*Runtime, error) {
+
+	if params.ID == "" {
+		var err error
+		params.ID, err = generateInstanceID()
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	loaded, err := loader.Filtered(params.Paths, params.Filter)
+	if err != nil {
+		return nil, err
+	}
+
+	store := inmem.New()
+
+	txn, err := store.NewTransaction(ctx, storage.WriteParams)
+	if err != nil {
+		return nil, err
+	}
+
+	// only register custom plugins if directory specified
+	if params.BuiltinDir != "" {
+		err = RegisterBuiltinsFromDir(params.BuiltinDir)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	if err := store.Write(ctx, txn, storage.AddOp, storage.Path{}, loaded.Documents); err != nil {
+		store.Abort(ctx, txn)
+		return nil, errors.Wrapf(err, "storage error")
+	}
+
+	if err := compileAndStoreInputs(ctx, store, txn, loaded.Modules, params.ErrorLimit); err != nil {
+		store.Abort(ctx, txn)
+		return nil, errors.Wrapf(err, "compile error")
+	}
+
+	if err := store.Commit(ctx, txn); err != nil {
+		return nil, errors.Wrapf(err, "storage error")
+	}
+
+	// register before init
+	if params.PluginDir != "" {
+		err = RegisterPluginsFromDir(params.PluginDir)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	m, plugins, err := initPlugins(params.ID, store, params.ConfigFile)
+	if err != nil {
+		return nil, err
+	}
+
+	var decisionLogger func(context.Context, *server.Info)
+
+	if p, ok := plugins["decision_logs"]; ok {
+		decisionLogger = p.(*logs.Plugin).Log
+
+		if params.DecisionIDFactory == nil {
+			params.DecisionIDFactory = generateDecisionID
+		}
+	}
+
+	rt := &Runtime{
+		Store:          store,
+		Manager:        m,
+		Params:         params,
+		decisionLogger: decisionLogger,
+	}
+
+	return rt, nil
 }
 
 // StartServer starts the runtime in server mode. This function will block the calling goroutine.
