+++ runtime/runtime_noplugins.go
@@ -6,77 +6,14 @@
 
 package runtime
 
-// Contains parts of the runtime package that do not use the plugin package
-import (
-	"context"
+// Contains parts of the runtime package that do not use the plugin package.
 
-	"github.com/pkg/errors"
+// RegisterBuiltinsFromDir is a no-op. Plugin loading is limited to linux/darwin + cgo platforms for the time being.
+func RegisterBuiltinsFromDir(dir string) error {
+	return nil
+}
 
-	"github.com/open-policy-agent/opa/loader"
-	"github.com/open-policy-agent/opa/plugins/logs"
-	"github.com/open-policy-agent/opa/server"
-	"github.com/open-policy-agent/opa/storage"
-	"github.com/open-policy-agent/opa/storage/inmem"
-)
-
-// NewRuntime returns a new Runtime object initialized with params.
-func NewRuntime(ctx context.Context, params Params) (*Runtime, error) {
-
-	if params.ID == "" {
-		var err error
-		params.ID, err = generateInstanceID()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	loaded, err := loader.Filtered(params.Paths, params.Filter)
-	if err != nil {
-		return nil, err
-	}
-
-	store := inmem.New()
-
-	txn, err := store.NewTransaction(ctx, storage.WriteParams)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := store.Write(ctx, txn, storage.AddOp, storage.Path{}, loaded.Documents); err != nil {
-		store.Abort(ctx, txn)
-		return nil, errors.Wrapf(err, "storage error")
-	}
-
-	if err := compileAndStoreInputs(ctx, store, txn, loaded.Modules, params.ErrorLimit); err != nil {
-		store.Abort(ctx, txn)
-		return nil, errors.Wrapf(err, "compile error")
-	}
-
-	if err := store.Commit(ctx, txn); err != nil {
-		return nil, errors.Wrapf(err, "storage error")
-	}
-
-	m, plugins, err := initPlugins(params.ID, store, params.ConfigFile)
-	if err != nil {
-		return nil, err
-	}
-
-	var decisionLogger func(context.Context, *server.Info)
-
-	if p, ok := plugins["decision_logs"]; ok {
-		decisionLogger = p.(*logs.Plugin).Log
-
-		if params.DecisionIDFactory == nil {
-			params.DecisionIDFactory = generateDecisionID
-		}
-	}
-
-	rt := &Runtime{
-		Store:          store,
-		Manager:        m,
-		Params:         params,
-		decisionLogger: decisionLogger,
-	}
-
-	return rt, nil
+// RegisterPluginsFromDir is a no-op. Plugin loading is limited to linux/darwin + cgo platforms for the time being.
+func RegisterPluginsFromDir(dir string) error {
+	return nil
 }
