+++ runtime/runtime_plugins.go
@@ -9,101 +9,15 @@ package runtime
 // Contains parts of the runtime package that use the plugin package
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"path/filepath"
 	"plugin"
 
-	"github.com/pkg/errors"
-
 	"github.com/open-policy-agent/opa/ast"
-	"github.com/open-policy-agent/opa/loader"
 	"github.com/open-policy-agent/opa/plugins"
-	"github.com/open-policy-agent/opa/plugins/logs"
-	"github.com/open-policy-agent/opa/server"
-	"github.com/open-policy-agent/opa/storage"
-	"github.com/open-policy-agent/opa/storage/inmem"
 	"github.com/open-policy-agent/opa/topdown"
 )
-
-// NewRuntime returns a new Runtime object initialized with params.
-func NewRuntime(ctx context.Context, params Params) (*Runtime, error) {
-
-	if params.ID == "" {
-		var err error
-		params.ID, err = generateInstanceID()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	loaded, err := loader.Filtered(params.Paths, params.Filter)
-	if err != nil {
-		return nil, err
-	}
-
-	store := inmem.New()
-
-	txn, err := store.NewTransaction(ctx, storage.WriteParams)
-	if err != nil {
-		return nil, err
-	}
-
-	// only register custom plugins if directory specified
-	if params.BuiltinDir != "" {
-		err = RegisterBuiltinsFromDir(params.BuiltinDir)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if err := store.Write(ctx, txn, storage.AddOp, storage.Path{}, loaded.Documents); err != nil {
-		store.Abort(ctx, txn)
-		return nil, errors.Wrapf(err, "storage error")
-	}
-
-	if err := compileAndStoreInputs(ctx, store, txn, loaded.Modules, params.ErrorLimit); err != nil {
-		store.Abort(ctx, txn)
-		return nil, errors.Wrapf(err, "compile error")
-	}
-
-	if err := store.Commit(ctx, txn); err != nil {
-		return nil, errors.Wrapf(err, "storage error")
-	}
-
-	// register before init
-	if params.PluginDir != "" {
-		err = RegisterPluginsFromDir(params.PluginDir)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	m, plugins, err := initPlugins(params.ID, store, params.ConfigFile)
-	if err != nil {
-		return nil, err
-	}
-
-	var decisionLogger func(context.Context, *server.Info)
-
-	if p, ok := plugins["decision_logs"]; ok {
-		decisionLogger = p.(*logs.Plugin).Log
-
-		if params.DecisionIDFactory == nil {
-			params.DecisionIDFactory = generateDecisionID
-		}
-	}
-
-	rt := &Runtime{
-		Store:          store,
-		Manager:        m,
-		Params:         params,
-		decisionLogger: decisionLogger,
-	}
-
-	return rt, nil
-}
 
 // RegisterBuiltinsFromDir recursively loads all custom builtins into OPA from dir. This function is idempotent.
 func RegisterBuiltinsFromDir(dir string) error {
